pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autojob-bot"
version = "0.1.0"
description = "Autonomous job application assistant"
authors = [{ name = "autojob-bot" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "alembic",
    "pydantic-settings",
    "redis",
    "celery[redis]",
    "python-dotenv",
    "httpx",
    "beautifulsoup4",
    "lxml",
    "feedparser",
    "structlog",
    "numpy",
    "scikit-learn",
    "email-validator",
    "python-dateutil",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "isort",
    "pytest",
    "pytest-asyncio",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "B", "UP", "S", "N", "ANN", "COM", "Q", "W"]
ignore = ["ANN101", "ANN102"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = ["tests"]
